Last commit of the day.
Changed the IsModifiable to it's own sub thinking it might be cleaner and perhaps work but it still thinks everything is ReadWrite.

I think it could be something to do with the fact that if I open a file the addin automatically reads iproperties and sets the file as editable? Not sure if that's the case though.